del: 기존 예외 코드 enum 삭제(#11)
