+++ src/main/java/com/bid/auction/global/enums/statuscode/error/CommonErrorStatus.java
@@ -1,0 +1,42 @@
+package com.bid.auction.global.enums.statuscode.error;
+
+import org.springframework.http.HttpStatus;
+
+import com.bid.auction.global.enums.statuscode.BaseCode;
+
+public enum CommonErrorStatus implements BaseCode {
+	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON500", "서버 에러, 관리자에게 문의 바랍니다."),
+	BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON400", "잘못된 요청입니다."),
+	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "COMMON401", "인증이 필요합니다."),
+	FORBIDDEN(HttpStatus.FORBIDDEN, "COMMON403", "금지된 요청입니다.");
+
+	private final HttpStatus httpStatus;
+	private final String code;
+	private final String message;
+
+	CommonErrorStatus(HttpStatus httpStatus, String code, String message) {
+		this.httpStatus = httpStatus;
+		this.code = code;
+		this.message = message;
+	}
+
+	@Override
+	public String getCode() {
+		return code;
+	}
+
+	@Override
+	public String getMessage() {
+		return message;
+	}
+
+	@Override
+	public HttpStatus getStatus() {
+		return httpStatus;
+	}
+
+	@Override
+	public Integer getStatusValue() {
+		return httpStatus.value();
+	}
+}
