+++ src/main/java/com/bid/auction/global/enums/statuscode/error/RefundErrorStatus.java
@@ -1,0 +1,39 @@
+package com.bid.auction.global.enums.statuscode.error;
+
+import org.springframework.http.HttpStatus;
+
+import com.bid.auction.global.enums.statuscode.BaseCode;
+
+public enum RefundErrorStatus implements BaseCode {
+	REFUND_REQUEST_NOT_FOUND(HttpStatus.BAD_REQUEST, "REFUND_REQUEST4001", "해당 환불 요청이 존재하지 않습니다");
+
+	private final HttpStatus httpStatus;
+	private final String code;
+	private final String message;
+
+	RefundErrorStatus(HttpStatus httpStatus, String code, String message) {
+		this.httpStatus = httpStatus;
+		this.code = code;
+		this.message = message;
+	}
+
+	@Override
+	public String getCode() {
+		return code;
+	}
+
+	@Override
+	public String getMessage() {
+		return message;
+	}
+
+	@Override
+	public HttpStatus getStatus() {
+		return httpStatus;
+	}
+
+	@Override
+	public Integer getStatusValue() {
+		return httpStatus.value();
+	}
+}
